requirements.txt
# GridGrab Backend Dependencies
# Python 3.10+ required

# Core Django framework
Django>=4.2,<5.0

# Django REST Framework for API endpoints
djangorestframework>=3.14.0

# CORS headers to allow frontend requests
django-cors-headers>=4.3.0

# Django Channels for WebSocket support
channels>=4.0.0

# Redis channel layer for production (optional for dev - uses InMemoryChannelLayer locally)
channels-redis>=4.2.0

# ASGI server for running Django with WebSocket support
daphne>=4.0.0

# Additional dependencies (included with Django but listed for clarity)
# asgiref - included with Django
# sqlparse - included with Django
